refactor: my-schedule 베럴패턴적용
