+++ src/features/my-schedule/api/delete-schedule.API.ts
