+++ src/features/my-schedule/hooks/use-date-schedules.ts
@@ -3,7 +3,8 @@ import { useSuspenseQuery } from '@tanstack/react-query'
 import { groupByDate } from '@/entities/schedule'
 import { scheduleQueryKeys } from '@/entities/schedule/models/schedule.query'
 import { useUserStore } from '@/entities/user/models/use-user-store'
-import { getMySchedule } from '@/features/my-schedule'
+
+import { getMySchedule } from '../api'
 
 export function useDateSchedules() {
   const userId = useUserStore((state) => state.user?.userId)
