+++ src/features/my-schedule/hooks/use-delete-schedule.ts
@@ -3,9 +3,8 @@ import { useMutation, useQueryClient } from '@tanstack/react-query'
 import { scheduleQueryKeys } from '@/entities/schedule/models/schedule.query'
 import { useUserStore } from '@/entities/user/models/use-user-store'
 import { toast } from '@/shared/ui/toast'
-import { devLog } from '@/shared/utils/dev-log'
 
-import deleteSchedule from '../detailed-schedule/api/delete-schedule.API'
+import deleteSchedule from '../api/delete-schedule.API'
 
 export default function useDeleteSchedule() {
   const userId = useUserStore((state) => state.user?.userId)
@@ -23,7 +22,7 @@ export default function useDeleteSchedule() {
     },
     onError: (error) => {
       toast.error('일정 삭제가 실패했습니다')
-      devLog('log', '일정삭제실패', error)
+      throw error
     },
   })
