--- src/features/my-schedule/index.ts
@@ -1,12 +1,0 @@
-// API
-export * from './api/my-schedule.API'
-
-// UI
-export { default as MyCalendar } from './my-calendar/my-calendar'
-export { default as MyDetailedScheduleList } from './detailed-schedule/ui/my-detailed-schedule-list'
-export { default as MyDetailedSchedule } from './detailed-schedule/ui/my-detailed-schedule'
-export { default as EditSchduleDropDown } from './detailed-schedule/ui/edit-schdule-dropdown'
-export { default as DeleteConfirmModal } from './detailed-schedule/ui/delete-confirm-modal'
-
-//hook
-export * from './hooks/use-date-schedules'
