+++ src/features/my-schedule/ui/delete-confirm-modal.tsx
@@ -1,6 +1,6 @@
-import { Modal, ModalContent, ModalOverlay, ModalPortal } from '../../../../shared/ui/modal'
-import Text from '../../../../shared/ui/text/text'
-import useDeleteSchedule from '../../hooks/use-delete-schedule'
+import { Modal, ModalContent, ModalOverlay, ModalPortal, Text } from '@/shared/ui'
+
+import useDeleteSchedule from '../hooks/use-delete-schedule'
 
 interface Props {
   isOpen: boolean
