+++ src/features/my-schedule/ui/my-detailed-schedule.tsx
